Move app path utilities into `files.hpp/.cpp`
