+++ nano/core_test/processor_service.cpp
@@ -1,11 +1,11 @@
 #include <nano/lib/blocks.hpp>
+#include <nano/lib/files.hpp>
 #include <nano/lib/logging.hpp>
 #include <nano/lib/stats.hpp>
 #include <nano/lib/work.hpp>
 #include <nano/node/make_store.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/secure/ledger_set_any.hpp>
-#include <nano/secure/utility.hpp>
 #include <nano/store/component.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
