+++ nano/core_test/work_pool.cpp
@@ -1,5 +1,6 @@
 #include <nano/crypto_lib/random_pool.hpp>
 #include <nano/lib/blocks.hpp>
+#include <nano/lib/files.hpp>
 #include <nano/lib/logging.hpp>
 #include <nano/lib/timer.hpp>
 #include <nano/lib/work.hpp>
@@ -7,7 +8,6 @@
 #include <nano/node/openclconfig.hpp>
 #include <nano/node/openclwork.hpp>
 #include <nano/secure/common.hpp>
-#include <nano/secure/utility.hpp>
 
 #include <gtest/gtest.h>
