+++ nano/lib/plat/darwin/working.mm
@@ -1,4 +1,4 @@
-#include <nano/secure/working.hpp>
+#include <nano/lib/files.hpp>
 
 #include <Foundation/Foundation.h>
