+++ nano/nano_wallet/entry_com.cpp
@@ -1,9 +1,8 @@
 #include <nano/lib/errors.hpp>
+#include <nano/lib/files.hpp>
 #include <nano/lib/utility.hpp>
 #include <nano/node/cli.hpp>
 #include <nano/rpc/rpc.hpp>
-#include <nano/secure/utility.hpp>
-#include <nano/secure/working.hpp>
 
 #include <boost/format.hpp>
 #include <boost/make_shared.hpp>
