--- nano/secure/working.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <nano/secure/common.hpp>
-
-namespace nano
-{
-std::filesystem::path app_path_impl ();
-}
